Ajustada tabela de pesquisa. Adicionada a regra de remoção de rotas e horários de rota só por pelo usuário dono.

+++ src/main/java/business/RouteBC.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.inject.Inject;
 
 import persistence.RouteDAO;
+import br.gov.frameworkdemoiselle.security.AuthorizationException;
 import br.gov.frameworkdemoiselle.stereotype.BusinessController;
 import br.gov.frameworkdemoiselle.transaction.Transactional;
 import entity.Coordinate;
@@ -18,9 +19,12 @@ public class RouteBC {
 
 	@Inject
 	private RouteDAO routeDAO;
-	
+
 	@Inject
 	private ScheduleBC scheduleBC;
+
+	@Inject
+	private UserBC userBC;
 
 	public void insert(Route route, User user) throws Exception {
 		routeDAO.insert(route, user);
@@ -32,16 +36,25 @@ public class RouteBC {
 
 	@Transactional
 	public void delete(Route route) throws Exception {
-		scheduleBC.delete(route);
-		routeDAO.delete(route);
+
+		Route _route = routeDAO.load(route.getId());
+		User _user = userBC.getCurrentUser();
+
+		if (_route.getUser().equals(_user)) {
+			scheduleBC.delete(route);
+			routeDAO.delete(route);
+		} else {
+			throw new AuthorizationException("Não é possível remover rota associada a outro usuário");
+		}
 	}
 
 	public List<Route> find(User user) throws Exception {
 		return routeDAO.find(user);
 	}
 
-	public List<Route> find(Coordinate coord, Integer radius, User user, Weekday weekday, Time hourini, Time hourend) throws Exception {
-		
+	public List<Route> find(Coordinate coord, Integer radius, User user, Weekday weekday, Time hourini, Time hourend)
+			throws Exception {
+
 		boolean empty = true;
 		empty &= coord == null;
 		empty &= radius == null;
@@ -49,12 +62,12 @@ public class RouteBC {
 		empty &= weekday == null;
 		empty &= hourini == null;
 		empty &= hourend == null;
-		
-		if(empty) {
+
+		if (empty) {
 			throw new Exception("erro...");
 		}
-		
-		return routeDAO.find(coord, radius, user, weekday, hourini, hourend); 
+
+		return routeDAO.find(coord, radius, user, weekday, hourini, hourend);
 	}
-	
+
 }
